docs: update examples/flutter_weather to cubit ^0.2.0
