+++ examples/flutter_weather/lib/app/weather/cubit/weather_cubit.dart
@@ -1,22 +1,8 @@
 import 'package:cubit/cubit.dart';
-import 'package:flutter/foundation.dart';
-
 import 'package:flutter_weather/service/service.dart';
+import 'package:meta/meta.dart';
 
-@immutable
-abstract class WeatherState {}
-
-class WeatherInitial extends WeatherState {}
-
-class WeatherLoadInProgress extends WeatherState {}
-
-class WeatherLoadSuccess extends WeatherState {
-  WeatherLoadSuccess(this.weather);
-
-  final Weather weather;
-}
-
-class WeatherLoadFailure extends WeatherState {}
+part 'weather_state.dart';
 
 class WeatherCubit extends Cubit<WeatherState> {
   WeatherCubit(this._weatherService) : super(WeatherInitial());
