+++ examples/flutter_weather/lib/app/weather/cubit/weather_state.dart
@@ -1,0 +1,16 @@
+part of 'weather_cubit.dart';
+
+@immutable
+abstract class WeatherState {}
+
+class WeatherInitial extends WeatherState {}
+
+class WeatherLoadInProgress extends WeatherState {}
+
+class WeatherLoadSuccess extends WeatherState {
+  WeatherLoadSuccess(this.weather);
+
+  final Weather weather;
+}
+
+class WeatherLoadFailure extends WeatherState {}
