+++ examples/flutter_weather/lib/app/weather/weather.dart
@@ -1,2 +1,3 @@
+export 'cubit/weather_cubit.dart';
 export 'weather_page.dart';
 export 'widgets/widgets.dart';
