+++ examples/flutter_weather/pubspec.yaml
@@ -9,8 +9,10 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  cubit: ^0.0.8
-  flutter_cubit: ^0.0.10
+  cubit:
+    path: ../../packages/cubit
+  flutter_cubit:
+    path: ../../packages/flutter_cubit
   google_fonts: ^1.1.0
   http: ^0.12.1
   json_annotation: ^3.0.1
@@ -19,7 +21,8 @@ dev_dependencies:
   flutter_test:
     sdk: flutter
   build_runner: ^1.8.1
-  cubit_test: ^0.0.5
+  cubit_test:
+    path: ../../packages/cubit_test
   json_serializable: ^3.3.0
   matcher: ^0.12.6
   mockito: ^4.0.0
